// src/app/connectors/api/model/eventDescriptionResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EventResponse } from './eventResponse';


export interface EventDescriptionResponse { 
    eventId?: number;
    objectivesText?: string;
    eventName?: string;
    eventDescription?: string;
    event?: EventResponse;
}