// src/app/connectors/api/model/eventRequest.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EventRequest { 
    initDate?: string;
    endDate?: string;
    city?: string;
    state?: string;
    country?: string;
    userId?: number;
    lon?: number;
    lat?: number;
    eventName?: string;
    objectivesText?: string;
    eventDescription?: string;
}